Restructure core commands in a consistent manner
