+++ packages/core/lib/commands/compile/meta.js
@@ -1,0 +1,50 @@
+module.exports = {
+  command: "compile",
+  description: "Compile contract source files",
+  builder: {
+    all: {
+      type: "boolean",
+      default: false
+    },
+    compiler: {
+      type: "string",
+      default: null
+    },
+    list: {
+      type: "string"
+    },
+    help: {
+      type: "boolean",
+      default: "false"
+    }
+  },
+  help: {
+    usage: "truffle compile [--list <filter>] [--all] [--quiet]",
+    options: [
+      {
+        option: "--all",
+        description:
+          "Compile all contracts instead of only the contracts changed since last compile."
+      },
+      {
+        option: "--list <filter>",
+        description:
+          "List all recent stable releases from solc-bin.  If filter is specified then it will display only " +
+          "that\n                    type of release or docker tags. The filter parameter must be one of the following: " +
+          "prereleases,\n                    releases, latestRelease or docker."
+      },
+      {
+        option: "--compiler <compiler-name>",
+        description:
+          "Specify a single compiler to use (e.g. `--compiler=solc`). Specify `none` to skip compilation."
+      },
+      {
+        option: "--save-intermediate <output-file>",
+        internal: true,
+        description:
+          "Save the raw compiler results into <output-file>, overwriting any existing content."
+      }
+    ],
+    allowedGlobalOptions: ["config", "quiet"]
+  }
+};
