+++ packages/core/lib/commands/create/meta.js
@@ -1,0 +1,32 @@
+module.exports = {
+  command: "create",
+  description: "Helper to create new contracts, migrations and tests",
+  builder: {
+    all: {
+      type: "boolean",
+      default: false
+    },
+    force: {
+      type: "boolean",
+      default: false
+    }
+  },
+  help: {
+    usage: "truffle create <artifact_type> <ArtifactName>",
+    options: [
+      {
+        option: "<artifact_type>",
+        description:
+          "Create a new artifact where artifact_type is one of the following: " +
+          "contract, migration,\n                    test or all. The new artifact is created " +
+          "along with one (or all) of the following\n                    files: `contracts/ArtifactName.sol`, " +
+          "`migrations/####_artifact_name.js` or\n                    `tests/artifact_name.js`. (required)"
+      },
+      {
+        option: "<ArtifactName>",
+        description: "Name of new artifact. (required)"
+      }
+    ],
+    allowedGlobalOptions: []
+  }
+};
