+++ packages/core/lib/commands/db/commands/query/meta.js
@@ -1,0 +1,10 @@
+module.exports = {
+  command: "query",
+  description: "Query @truffle/db",
+  builder: {},
+  help: {
+    usage: "truffle db query <query>",
+    options: [],
+    allowedGlobalOptions: []
+  }
+};
