+++ packages/core/lib/commands/db/commands/serve/meta.js
@@ -1,0 +1,10 @@
+module.exports = {
+  command: "serve",
+  description: "Start Truffle's GraphQL UI playground",
+  builder: {},
+  help: {
+    usage: "truffle db serve",
+    options: [],
+    allowedGlobalOptions: []
+  }
+};
