+++ packages/core/lib/commands/deploy/index.js
@@ -1,15 +1,4 @@
-const migrate = require("../migrate");
-
-const command = {
-  command: "deploy",
-  description: "(alias for migrate)",
-  builder: migrate.builder,
-  help: {
-    usage:
-      "truffle deploy [--reset] [-f <number>] [--compile-all] [--verbose-rpc]",
-    options: migrate.help.options,
-    allowedGlobalOptions: ["network", "config"]
-  }
+module.exports = {
+  run: require("./run"),
+  meta: require("./meta")
 };
-
-module.exports = command;
