+++ packages/core/lib/commands/help/index.js
@@ -1,18 +1,4 @@
-const command = {
-  command: "help",
-  description:
-    "List all commands or provide information about a specific command",
-  help: {
-    usage: "truffle help [<command>]",
-    options: [
-      {
-        option: "<command>",
-        description: "Name of the command to display information for."
-      }
-    ],
-    allowedGlobalOptions: []
-  },
-  builder: {},
+module.exports = {
+  run: require("./run"),
+  meta: require("./meta")
 };
-
-module.exports = command;
