+++ packages/core/lib/commands/init/index.js
@@ -1,20 +1,4 @@
-const command = {
-  command: "init",
-  description: "Initialize new and empty Ethereum project",
-  builder: {},
-  help: {
-    usage: "truffle init [--force]",
-    options: [
-      {
-        option: "--force",
-        description:
-          "Initialize project in the current directory regardless of its " +
-          "state. Be careful, this\n                    will potentially overwrite files " +
-          "that exist in the directory.",
-      },
-    ],
-    allowedGlobalOptions: []
-  }
+module.exports = {
+  run: require("./run"),
+  meta: require("./meta")
 };
-
-module.exports = command;
