+++ packages/core/lib/commands/init/meta.js
@@ -1,0 +1,18 @@
+module.exports = {
+  command: "init",
+  description: "Initialize new and empty Ethereum project",
+  builder: {},
+  help: {
+    usage: "truffle init [--force]",
+    options: [
+      {
+        option: "--force",
+        description:
+          "Initialize project in the current directory regardless of its " +
+          "state. Be careful, this\n                    will potentially overwrite files " +
+          "that exist in the directory."
+      }
+    ],
+    allowedGlobalOptions: []
+  }
+};
