+++ packages/core/lib/commands/publish/meta.js
@@ -1,0 +1,10 @@
+module.exports = {
+  command: "publish",
+  description: "Publish a package to the Ethereum Package Registry",
+  builder: {},
+  help: {
+    usage: "truffle publish",
+    options: [],
+    allowedGlobalOptions: []
+  }
+};
