+++ packages/core/lib/commands/run/meta.js
@@ -1,0 +1,15 @@
+module.exports = {
+  command: "run",
+  description: "Run a third-party command",
+  builder: {},
+  help: {
+    usage: "truffle run [<command>]",
+    options: [
+      {
+        option: "<command>",
+        description: "Name of the third-party command to run."
+      }
+    ],
+    allowedGlobalOptions: []
+  }
+};
