+++ packages/core/lib/commands/version/index.js
@@ -1,12 +1,4 @@
-const command = {
-  command: "version",
-  description: "Show version number and exit",
-  builder: {},
-  help: {
-    usage: "truffle version",
-    options: [],
-    allowedGlobalOptions: []
-  }
+module.exports = {
+  run: require("./run"),
+  meta: require("./meta")
 };
-
-module.exports = command;
