+++ packages/core/lib/commands/version/meta.js
@@ -1,0 +1,10 @@
+module.exports = {
+  command: "version",
+  description: "Show version number and exit",
+  builder: {},
+  help: {
+    usage: "truffle version",
+    options: [],
+    allowedGlobalOptions: []
+  }
+};
