+++ packages/core/lib/commands/watch/index.js
@@ -1,13 +1,4 @@
-const command = {
-  command: "watch",
-  description:
-    "Watch filesystem for changes and rebuild the project automatically",
-  builder: {},
-  help: {
-    usage: "truffle watch",
-    options: [],
-    allowedGlobalOptions: []
-  }
+module.exports = {
+  run: require("./run"),
+  meta: require("./meta")
 };
-
-module.exports = command;
