+++ packages/core/lib/commands/watch/meta.js
@@ -1,0 +1,11 @@
+module.exports = {
+  command: "watch",
+  description:
+    "Watch filesystem for changes and rebuild the project automatically",
+  builder: {},
+  help: {
+    usage: "truffle watch",
+    options: [],
+    allowedGlobalOptions: []
+  }
+};
